requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
sqlalchemy==2.0.28
mysql-connector-python
Pyrebase4